move scss to 'styles' folders
