+++ app/src/components/searchWidget/date-input/styles/dateInput.scss
