+++ app/src/components/searchWidget/date-input/styles/datepicker.scss
