+++ app/src/components/searchWidget/styles/slider.scss
