+++ app/src/components/searchWidget/styles/widget.scss
